update windows tests and split sediment test in new file
